app/api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
firebase-admin
loguru
Faker
pytz
geopy
